fix(labor): days_worked as SUM of shift multipliers, not entry count (#51)
* fix(labor): days_worked as SUM of shift multipliers, not entry count

Previously days_worked was `COUNT(*)` over priced rows: a half-day
shift counted as 1 day, same as a full-day. A mix of 2 full + 2 half
shifts rendered as 4 days even though the worker actually delivered
3.0 priced days and was paid for 3.0 days. Reported on the Summary
tab as "Huấn shows 88 €/day average when the configured rate is
110 €/day" — the rate denominator (days_worked = 5) was wrong; the
truthful value is 4.0.

Switch the aggregation to SUM(shift_multiplier):
  - full → 1.0
  - half → 0.5
  - overtime → 1.5
  - supplement-only (shift_type IS NULL) → 0
This restores the invariant cost / days_worked == daily_rate for
every non-override row, and gives users a number that reconciles
with reality (e.g. 2.5 days for a 2-full + 1-half month).

Types follow the data:
  - ports.LaborSummaryRow.days_worked: int → Decimal
  - ports.MonthlyWorkerSubRow.days_worked: int → Decimal
  - ports.MonthlyLaborSummaryRow.total_days: int → Decimal
  - DTOs and Pydantic schemas: int → float
The PDF builder gains a small `_format_days` helper that drops
trailing zeros (5 instead of 5.0, 2,5 for 2.5) and the XLSX builder
needs no code change — float values land natively in Excel cells.

Test updates:
  - test_get_monthly_summary_excludes_supplement_only_from_totals:
    March bucket (2 full + 1 half) now asserts 2.5, not 3. The
    existing inline comment already anticipated this.
  - test_get_monthly_summary_nests_per_worker_breakdown:
    March's lone half-day worker now asserts 0.5, not 1.

* chore: bump version to 0.2.39

---------

+++ app/application/labor/get_labor_summary.py
@@ -13,7 +13,9 @@ from app.application.labor.ports import ILaborEntryRepository
 class WorkerCostSummary:
     worker_id: str
     worker_name: str
-    days_worked: int
+    # Fractional — see ports.LaborSummaryRow.days_worked. A mix of
+    # full + half shifts cleanly produces 2.5, not 3.
+    days_worked: float
     total_cost: float
     banked_hours: int
     bonus_full_days: int
@@ -24,7 +26,8 @@ class WorkerCostSummary:
 @dataclass
 class LaborSummaryResponse:
     rows: List[WorkerCostSummary]
-    total_days: int
+    # Sum of per-worker days_worked — also fractional.
+    total_days: float
     total_cost: float
     total_banked_hours: int
     total_bonus_days: float
@@ -74,7 +77,7 @@ class GetLaborSummaryUseCase:
                 WorkerCostSummary(
                     worker_id=str(row.worker_id),
                     worker_name=row.worker_name,
-                    days_worked=row.days_worked,
+                    days_worked=float(row.days_worked),
                     total_cost=float(total_cost_for_worker),
                     banked_hours=banked,
                     bonus_full_days=bonus_full,
@@ -87,7 +90,7 @@ class GetLaborSummaryUseCase:
             total_bonus_days += worker_bonus_days
             total_bonus_cost += bonus_cost
 
-        total_days = sum(r.days_worked for r in rows)
+        total_days = float(sum(r.days_worked for r in rows))
         total_cost = sum(r.total_cost for r in rows)
 
         return LaborSummaryResponse(
